requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0
Pillow==10.1.0
typing-extensions==4.8.0
